7-print_last_digit.c

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -6,7 +6,7 @@
  *
  * Return: The value of the last digit.
  */
-int print_last_digit(int n)
+int print_last_digit(long n)
 {
 	int last_digit;
